Use autoconf tools
I'm working with some other tools (fizmo-remglk, in particular) that
consumes this library, but it has to jump through some hoops in order to
find it.  While migrating to autoconf might be overkill jsut to solve
this problem, it also gives us easy installation in a standard way,
which used to be a post-build manual step.

Add .gitignore file so generated files aren't accidentally committed.

Add consfigure.ac to drive the autoconf process.

Replace Makefile with Makefile.am (which is consumed by
autoconf/configure to generate the final Makefile)

Add remglk.pc.in for pkg-config metadata generation.

Tweak README.txt to include the new build/install/how-to-use
instructions.